requirements.txt
aiohttp==3.6.2
argcomplete==1.12.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
BASC-py4chan==0.6.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord.py==1.4.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
meson==0.55.0
multidict==4.7.6
mypy==0.782
mypy-extensions==0.4.3
nekos.py==1.0.3
pipx==0.15.4.0
praw==7.1.0
prawcore==1.5.0
protobuf==3.12.4
pycodestyle==2.6.0
pylint==2.5.3
requests==2.24.0
six==1.15.0
switch==1.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
update-checker==0.18.0
urllib3==1.25.10
userpath==1.4.1
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.5.1
